* Code cleanup , bug fix , expand buttonadded
