+++ src/bento/announcementPageData.js
@@ -1,0 +1,59 @@
+
+import fig1 from "../assets/announcement/fig1.svg";
+import fig2 from "../assets/announcement/fig2.svg";
+import fig3 from "../assets/announcement/fig3.svg";
+import fig4 from "../assets/announcement/fig4.svg";
+import fig6 from "../assets/announcement/fig6.svg";
+import fig7 from "../assets/announcement/fig7.svg";
+
+
+export const announcementPageData = [
+    {
+        "title": "Resource Update",
+        "timestamp": "October 30, 2024",
+        "image": fig1,
+        "type": 1,
+        "alt": "Megaphone",
+        "verbiage": "We are excited to announce that Release 4 of the C3DC application is now available!  This release introduces new sections under the About page, including an Announcement page and Release Notes, to enhance the user experience and provide more comprehensive data updates."
+    },
+    {
+        "title": "Cohort Selector",
+        "timestamp": "October 30, 2024",
+        "image": fig2,
+        "type": 1,
+        "alt": "A woman wearing glasses sitting in front of a computer",
+        "verbiage": "This new feature allows users to select up to 10 cohort groups (participants), generate and download clinical and manifest files in CSV or JSON format, and utilize them on other platforms."
+    },
+    {
+        "title": "Cohort Analyzer",
+        "timestamp": "October 30, 2024",
+        "image": fig4,
+        "type": 1,
+        "alt": "A laptop screen displaying a progress bar with the text 'Updating' shown on the screen.",
+        "verbiage": "This feature will be available in the upcoming January release. It will enable users to compare up to 3 cohorts and identify shared participants and diseases."
+    },
+    {
+        "title": "Explore Page",
+        "timestamp": "October 30, 2024",
+        "image": fig3,
+        "type": 1,
+        "alt": "A top-down view of a laptop, with a person typing on the keyboard",
+        "verbiage": "In this release, we have reorganized the facet search, and the Ethnicity widget has been replaced with a new Treatment widget."
+    },
+    {
+        "title": "Datasets Update",
+        "timestamp": "October 30, 2024",
+        "image": fig6,
+        "type": 2,
+        "alt": "An animated laptop surrounded by animated objects, including a server, gears, and a magnifying glass.",
+        "verbiage": "In this release, we've added new datasets to the C3DC explore page, including TARGET data for Osteosarcoma (OS), Kidney, CCSK, AML, Wilms Tumor (WT), ALL Expansion Phase 2, Rhabdoid Tumor (RT), Cancer Model Systems (MDLS), and ALL Pilot Phase 1, along with new CCDI datasets."
+    },
+    {
+        "title": "Data Model Update",
+        "timestamp": "July 24, 2024",
+        "image": fig7,
+        "type": 2,
+        "alt": "A web-like design connecting various icons, including people icons and profile icons.",
+        "verbiage": "In this release, we are introducing enhancements to our data model. Specifically, we have added two new nodes: treatment and treatment response. These additions will allow for more detailed tracking and analysis of treatment regimens and their outcomes, providing valuable insights for researchers and clinicians."
+    }
+];
