--- src/bento/announcmentPageData.js
@@ -1,97 +1,0 @@
-export const announcementPageData = [
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Data Update: Lorem ipsum odor amet consectetuer em ipsum",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        }
-        ,  {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },  {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },  {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024 ",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },  {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024",
-            image: "",
-            type: 1,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        },
-        {
-            title: "Application Update: Lorem ipsum odor amet consectetuer adipiscing elit",
-            timestamp: "October 31, 2024",
-            image: "",
-            type: 2,
-            verbiage: "we are excited to announce that release 3 of the C3DC application is available now. in this release, we will introduce additional pages such as data model, resources, and studies page to the site to enhance user experiance and provide more comprehensive data access. Lorem ipsum odor amet, constref adipci elit, auguge virvere elementum interdum tactic manm. Ac integer pellentsque him"
-        }
-];
